Move StreamConfig from SearchConfig to UpdaterConfig; remove unused DeleteConfig and DeletionConfig.
