+++ services/thingsearch/common/src/main/java/org/eclipse/ditto/services/thingsearch/common/config/SearchConfig.java
@@ -43,33 +43,11 @@ public interface SearchConfig extends ServiceSpecificConfig, WithHealthCheckConf
     String getQueryValidatorImplementation();
 
     /**
-     * Returns the configuration settings of the "delete" section.
-     *
-     * @return the config.
-     */
-    DeleteConfig getDeleteConfig();
-
-    /**
-     * Returns the configuration settings for the physical deletion of thing entities that are marked as
-     * {@code "__deleted"}.
-     *
-     * @return the config.
-     */
-    DeletionConfig getDeletionConfig();
-
-    /**
      * Returns the configuration settings for the search updating functionality.
      *
      * @return the config.
      */
     UpdaterConfig getUpdaterConfig();
-
-    /**
-     * Returns the configuration settings
-     *
-     * @return the config.
-     */
-    StreamConfig getStreamConfig();
 
     /**
      * An enumeration of the known config path expressions and their associated default values for SearchConfig.
