+++ services/thingsearch/starter/src/main/resources/things-search.conf
@@ -43,29 +43,10 @@ ditto {
     query-criteria-validator.implementation = ${?QUERY_CRITERIA_VALIDATOR_IMPLEMENTATION} // TODO ff also rename
     mongo-hints-by-namespace = ${?MONGO_HINTS_BY_NAMESPACE}
 
-    delete {
-      event = true
-      event = ${?THINGS_SEARCH_DELETE_EVENT}
-      namespace = true
-      namespace = ${?THINGS_SEARCH_DELETE_NAMESPACE}
-    }
-
     index-initialization {
       #indices should be created within this application
       enabled = true
       enabled = ${?INDEX_INITIALIZATION_ENABLED}
-    }
-
-    # configuration regarding physical deletion of "__deleted" Things from "thingEntities" collection
-    deletion {
-      enabled = true
-
-      # delete "__deleted" Things older than:
-      deletion-age = 3d
-      # run each:
-      run-interval = 24h
-      # the first run should be at (UTC time):
-      first-interval-hour = 21 # 21:00 UTC
     }
 
     updater {
